separate masthead component and fix test
